ACPX extension: split ACP stream parser and test fixtures
